Consolidate parallel pool internals into pool.go
